+++ src/settings/settings-tab.ts
@@ -20,6 +20,7 @@ export type { SettingsTabName } from "./tab-names";
 
 // Tab renderer imports
 import { renderGeneralSettingsTab } from "./tabs/general-settings-tab";
+import { renderFreshRssSettingsTab } from "./tabs/freshrss-settings-tab";
 import { renderStorageSettingsTab } from "./tabs/storage-settings-tab";
 import { renderDisplaySettingsTab } from "./tabs/display-settings-tab";
 import { renderSidebarSettingsTab } from "./tabs/sidebar-settings-tab";
@@ -95,6 +96,10 @@ export class RssDashboardSettingTab extends PluginSettingTab {
     switch (this.currentTab) {
       case "General":
         renderGeneralSettingsTab(tabContent, this.plugin);
+        this.pendingSection = null;
+        break;
+      case "FreshRSS":
+        renderFreshRssSettingsTab(tabContent, this.plugin);
         this.pendingSection = null;
         break;
       case "Storage":